Make the TodoItem reads project but not able to update.

+++ csharp-example/TodoApi/TodoApi/Models/TodoItem.cs
@@ -12,7 +12,7 @@ namespace TodoApi.Models
 
         public string? Secret { get; set; }
 
-        public long? ProjectId { get; set; }
+        public int? ProjectId { get; set; }
 
         [ForeignKey("ProjectId")]
         public Project? Project { get; set; }
